requirements.txt
aioredis==1.3.1
asgiref==3.4.1
astroid==2.3.3
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==4.0.2
django-admin-interface==0.21.0
django-allauth==0.51.0
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.2
django-ckeditor==6.2.0
django-cleanup==5.2.0
django-colorfield==0.7.2
-e git://github.com/maraujop/django-crispy-forms.git@db00d2fc4766ac445d4fd74ceba2cb564c0eb7d6#egg=django_crispy_forms
django-datetime-widget-fa==0.10.0
django-datetime-widget2==0.9.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-recaptcha==3.0.0
gevent==21.12.0
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.3
oauthlib==3.2.2
Pillow==9.0.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.0.0
python3-openid==3.2.0
pytz==2022.1
requests==2.22.0
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.13.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
sslcommerz-python==0.0.7
swapper==1.3.0
toml==0.10.2
Twisted==22.2.0
txaio==22.2.1
typed-ast==1.4.0
typing-extensions==4.1.1
urllib3==1.25.7
websocket==0.2.1
wrapt==1.11.2
zope.event==4.5.0
zope.interface==5.4.0
